+++ ui/src/pages/chat/split-layout.test.ts
@@ -1,6 +1,7 @@
 // @vitest-environment node
 import { describe, expect, it } from "vitest";
 import { normalizeChatSplitLayout } from "./split-layout-persistence.ts";
+import type { ChatSplitLayout } from "./split-layout-types.ts";
 import {
   applyUiCommandToSplitLayout,
   closePane,
@@ -11,7 +12,6 @@ import {
   resizePanes,
   setActivePane,
   setPaneSession,
-  type ChatSplitLayout,
 } from "./split-layout.ts";
 
 function createSinglePaneLayout(sessionKey: string): ChatSplitLayout {
